Add a `Model` type for the solver
If all parameters of a model are independent of beta, adding interactions is
relatively easy; however, this is in general not the case.

Parameters such as masses and widths will in general have a temperature
dependence, and various coupling strengths can be modified by RGEs.

The computation of each of these is done once at the generation of the context
and can be accessed repeatedly for the various interactions.

Signed-off-by: JP-Ellis <josh@jpellis.me>