Another rewrite of interactions:
Asymmetry is now expected to be given as δM^2 and calculates δγ.  This is now
better documented in the documentation as well.

Change the rate so that it has the symmetric and asymmetric rates, instead of
treating the forward and backward rates separately.

Also minor refactoring to use 1-indexing instead of 0-indexing.